delete the cast of my char in integer

+++ Jour01/Job03/char_is_digit.c
@@ -2,12 +2,9 @@
 // Created by corentin on 27/11/23.
 //
 
-
-#include <stdio.h>
-
 int char_is_digit(char c)
 {
-    if((int)c < 48 || (int)c > 57)
+    if(c < 48 || c > 57)
     {
         return 0;
     }
