+++ Jour01/Job03/str_is_digit.c
@@ -2,7 +2,6 @@
 // Created by corentin on 27/11/23.
 //
 
-#include <stdio.h>
 
 int str_is_digit(char *str)
 {
@@ -10,7 +9,7 @@ int str_is_digit(char *str)
 
     while(str[i] != 0)
     {
-        if((int)str[i] > 57 || (int)str[i] < 48)
+        if(str[i] > 57 || str[i] < 48)
         {
             return 0;
         }
